Replace nom::IResult with std::Result
Using nom::internal::IResult in the crate's public API required users to depend
on nom as well for error handling. This dependency is removed by returning the
std Result in the public API. For now the 'nom parser functions' are preserved
as the internal_ functions so we can make them public again if required.